Exercício 02 da aula 06

+++ Exercício 01 da aula 06.py
@@ -1,8 +1,9 @@
+#Receber 10 notas e tirar a média usando while
 n = 10
 soma = 0
 i = 1
 while i <= 10:
-    num = int(input("Digite um número: "))
+    num = float(input("Digite um número: "))
     soma += num
     i+=1
 media = soma/10
